requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
supabase==2.4.6
trafilatura==1.9.0
apscheduler==3.10.4
python-dotenv==1.0.1
